Avoid adding annotations to the wrong item when extracting
Previously, `ExtractAnnotations` would add annotations to the wrong
`<item>` element when writing the annotations. That was because it
would use `item != prev` to determine if the current open `<item>`
element needed closing before adding the `<annotation>`. In some cases
that check would produce the wrong result and treat `item` as equal to
`prev` even if they were different.

That would occur because:
1. `PsiMethodItem` defines `equals` in terms of the underlying
   `PsiMethod`.
2. When inheriting a method from a hidden class to a publicly visible
   class while a new `PsiMethodItem` was created it would reused the
   `PsiMethod` from the hidden class.
3. Inheriting the same hidden `PsiMethodItem` into two separate classes
   would create two `PsiMethodItem`s that share the same `PsiMethod`
   and so are equal even though they in different classes.
4. Sometimes those two methods would be next to each other in the list
   of `Item`/`AnnotationHolder` pairs.

If all of those conditions were met then when the second method in the
list was processed it would be treated as if it was equal to the first
method and so the `<item>` element would not be closed. That would
result in the annotation from the second method being added to the
`<item>` for the first method.

The fix was to compare signatures and not `Item`s. Updating the
`PsiMethodItem.equals(...)` to work better could have repercussions
elsewhere in Metalava so will be done as a follow-up to avoid delaying
this fix.

Bug: 329116156
Test: ./gradlew
Change-Id: Iba1575fb8dde5d81968ef2a6b3c6e9569e477091