Add ExtractAnnotations test for inherited methods
Reproduces the downstream issue where when a hidden method with a type
def annotation (but it applies to any annotation that is extracted) is
inherited into two (or more classes) and the two method are adjacent to
each other in the package annotation list (when sorted by signature)
the annotation is repeated on the first method and ignored on the
second and subsequent methods.

This test shows the current, albeit broken, behavior of Psi as it makes
it easier to review that this is reproducing the downstream issue. It
will be fixed in the follow up change. The test breaks in Turbine too
but for different reasons.

Bug: 329116156
Test: ./gradlew
Change-Id: Ie40ba39fce4f8feabf4a77726fd2e8d1010a2723