Fix issue with trailing whitespace in annotation attributes
Previously, when parsing an annotation that had whitespace after the
last attribute value and before the closing `)`,
`DefaultAnnotationAttribute.createList(...)` would erroneously add a
`value` attribute for the whole attribute source. That does not cause
issues in practice as the method is only used by API signature files
and they never add white space at the end of the attribute list.
However, it does break some tests which do have such whitespace in the
API signature file inputs. This change fixes it as it makes the code
more robust.

Bug: 325691395
Test: ./gradlew
Change-Id: If0bdfc9a38701729847e16e498e13edc1c3d5705