Treat revert annotations as hide annotations
Previously, `DefaultAnnotationManager.hasHideAnnotations(ModifierList)`
would always return `false` if the configured list of hide annotations
was empty. That was an optimization to avoid checking if any of the
modifier's `AnnotationItem`s returned `true` from `showability.hide()`
as hide annotations are hardly ever used. However, revert annotations
can also cause `showability.hide()` to return `true`, i.e. when
reverting a change that added a new API. So, that optimization was
preventing some newly added API items from being hidden.

The API items that were not being hidden correctly were fields in
interfaces. The code relies on `FieldItem.originallyHidden` to be set
to `true` in order for the fields to be correctly hidden. That can
happen if the field is explicitly hidden using something like `@hide`
or if `hasHideAnnotations(ModifierList)` returns `true`. However, the
optimization caused that to return `false` which meant
`FieldItem.originallyHidden` was set to `false` and so the field was
not hidden.

This change fixes that by only optimizing if the configured list of
revert annotations is also empty.

Bug: 332611311
Test: ./gradlew
Change-Id: I3b4df2c18b801f303cb2739b60349a1549cc3f7b