Support controlling Item.emit per signature file
The `Item.emit` property is one of the many aspects of an `Item` that
can affect whether it is output. An `Item` will not be output if `emit`
is set to `false`. It does not just affect what Metalava writes in API
signature files or stubs. It also affects other behavior. e.g.
Compatibility issues will not be reported if the `Item` from the
previously released API has `Item.emit` set to `false`.

For `Item`s created from signature files the `emit` property is
effectively an indication whether the `Item` is related to the current
API that Metalava is generation. That is because:
1. Metalava currently only uses `TextCodebase`s that correspond to the
   API surface that is being generated, e.g. if the system API is being
   created then only the `TextCodebase` from the previously released
   system API delta will be used.
2. The `emit` property is `true` for a `TextItem` created from a
   signature file and `false` for any stubs.

Unfortunately, to correctly support `--revert-annotation` Metalava
needs to be able to access the whole previously released API for the
API being generated, e.g. the previously released public API plus the
system API delta. That means it cannot just use the `TextCodebase` for
the API surface to be generated.

Providing access to the whole API when needed while maintaining the
current behavior elsewhere, requires setting `Item.emit` to `true`
for `Item`s loaded from the signature file for the current API while
setting it to `false` for `Item`s loaded from other signature files.

This change adds the `SignatureFile.forCurrentApiSurface` property
which determines whether the file is for the current API surface or
not. `ApiFile` then uses that to determine the value of `Item.emit`
for `Item`s it creates from that file.

Follow-up changes will make use of the new property to improve the
handling of `--revert-annotation`.

Bug: 319874764
Test: ./gradlew
Change-Id: I3f94bf63d9fd97c6e0fd4a40fcb2162441d891e8