Add SignatureFile wrapper around File for ApiFile
A follow-up change needs to pass additional information alongside the
`File` to `ApiFile` when parsing signature files that are used by the
compatibility check. This change adds a `SignatureFile` wrapper and
updates `ApiFile` and calling code to use it.

An additional `ApiFile.parseApi(List<File>)` method was added, and
annotated with `MetalavaApi` as it is needed for some uses from
outside Metalava.

Bug: 319874764
Test: ./gradlew
Change-Id: I538fb0e8df67f23e9cce559ba887b7dda5819e85