Remove codebaseValidator from ActionContext.loadFromJarFile(...)
The `codebaseValidator` parameter was added by change
https://r.android.com/2745410 to abstract away dependencies on the
`options.nullabilityAnnotationsValidator`. That in turn was added by
https://r.android.com/826544 to support validating nullability
annotations from a list. However, that logic is not actually needed
when loading a jar as we never load a jar when the
`options.nullabilityAnnotationsValidator` is available as that is only
used in `art-module-public-api-stubs-nullability-validation` and that
does not load a jar. Also, there is no point in validating nullability
annotations on a jar as we do not create APIs from jars except in some
limited circumstances and in that case the jars are stub jars created
by Metalava. So, if needed, Metalava would already have run nullability
checks on them.

Bug: 335616501
Test: ./gradlew
Change-Id: I843f854bf7223cd2de47b4c260e280e61ccf0716