Prevent "sdks" containing SDK extension later than "since"
Previously, it was possible to have something like this in an
`api-versions.xml` file:
```
   <... since="31" sdks="34:1,0:31">
```

Notice that it is listed as being released in SDK extension 1 of
dessert release 34 (T) even though it has been released as part of an
earlier dessert release. That causes issues for Android Lint as
outlined in the previous change. This fixes that by preventing the
addition of any SDK extension for a dessert release after the dessert
release in which the API was released.

It has to add special support for dessert release independent sdks,
i.e. those whose id is greater than or equal to 1,000,000, e.g.
`AD_SERVICES`. Otherwise, they would always be removed.

Bug: 336993217
Test: ./gradlew
Change-Id: I71fd67727a3753339da931210b46e42f1313947c