Avoid `Extract System Api` test depending on deprecated behavior
Previously, the `Extract System Api` included checks for entries in
the `api-versions.xml` files containing something like this:
```
    .... since=\"31\" sdks=\"33:1,0:31\"/>
```

i.e. where the `sdks` attribute includes extensions like `33:1` that
are greater than the `since` value, i.e. the first dessert release in
which the API was released. That causes problems for Android Lint and
will be changed. This change fixes the test to check for other APIs
instead to maintain the existing test coverage when a follow-up change
fixes the incorrect behavior.

Bug: 336993217
Test: ./gradlew

    Bug: 336993217
    Test: ./gradlew

Change-Id: I458a9c395ba577af0f350d638ea55acccf112d2b