Update text component type nullability from context
Needed for annotation members, which have implicitly non-null array component types.

The test for this case uncovered an issue with parsing when `kotlin-name-type-order=yes` where an API name might be parsed as a modifier instead of a name (b/338434680), the test is updated to work around that for now so the annotation array component nullability can be tested.

Bug: 325454211
Test: removed previously failing test from baseline

Change-Id: Icec4e167c442fbaac52d4712b8f46ba25bd0dd5a