Simplify text nullability synchronization
Previously a TextItem's modifiers were created with no annotations, then after the type for the item was created a nullability annotation might be added to the list of annotations for the item, and then the annotations were added to the modifiers. This simplifies things by creating the modifiers with the initial list of annotations and then adding a nullability annotation directly to the modifiers if needed based on the type.

Bug: 332507644
Test: ./gradlew test

Change-Id: Ib21d8572759be668b68f7a522a2299ef8684044e