Only skip printing object bounds if they have default nullability
During signature file parsing, when Kotlin style nulls are used in a signature file, an unbounded wilcard (?) is interpreted as having a non-null object bound, and when Kotlin style nulls are not used, an unbounded wildcard is interpreted as having a platform object bound. This is consistent with the "default" nullability of types for each format.

This updates the printing of types to follow the same convention, as otherwise writing and then reading a signature file will lose nullability information.

Bug: 332507644
Test: Added type string tests for all object bound cases

Change-Id: I2d7ab1040cad98eb1b70b67c0424f1be7ced2a4f