Add test showing findCorrespondingItemIn(...) does not find super methods
A follow-up change needs to use `Item.findCorrespondingItemIn(...)` to
find a previously released `Item`, if any. It will be used to try and
replicate the existing behavior of `CodebaseComparator`. That has
special support for handling methods which may have been elided from
the API signature file. Something similar is needed in
`Item.findCorrespondingItemIn(...)`. This change adds a test to show
the existing behavior. A follow-up change will enhance the method to
optionally support finding super methods.

The test uses nested calls to `runCodebaseTest(...)` which causes a
problem for the `psi` model as it does not support creating two
`PsiEnvironmentManager`s at the same time; it will fail when trying to
clean them up. So, it was necessary to fix that as part of this change.

Bug: 311690008
Test: ./gradlew
Change-Id: I12f21ace3e31926dde37c02d0e7379ab8b82ae81