Finish work on DefaultReporter's maximumSeverity parameter
Previously, a `maximumSeverity` parameter was added to the
`Reporter.report(...Reportable...)` method for use by `ApiLint` to
filter out issues on previously released APIs to replicate the
legacy behavior of `ApiLint` when it used `CodebaseComparator`.
Unfortunately, no tests were added and the `report(...File...)`
method was inconsistent as it did not have a `maximumSeverity`
parameter.

This change adds tests and makes the `report(...File...)` method
consistent.

Bug: 311690008
Test: ./gradlew
Change-Id: I3896106cee822442394d73aab356add8447a7332