Add tests for ApiLint filtering
Change https://r.android.com/3074232 replaced the use of
`CodebaseComparator` to restrict lint errors to only new APIs with a
new filtering mechanism that will make it easier to customize the
behavior for different issues and report warnings on new members of
existing classes.

Unfortunately, the behavior of `ApiLint` that relied on the effect
of `CodebaseComparator` was not well tested. This change adds some
additional tests to cover the filtering behavior.

Bug: 311690008
Test: ./gradlew
Change-Id: I8773d7dbcd4d8114a47b648d0972e3f4ae961c08