Update nullness compatibility check to use types
The check was previously based on the item annotations. This updates the error message in the removing nullability case, but that shouldn't impact any existing suppressions since they are based on issue ID.

As shown by some of the updated tests, this error might get triggered when a type completely changes.

Bug: 332507644
Test: Update existing nullness compat tests
Change-Id: Iffa929346af3875494da89804f44831f531c6195