Remove unnecessary API_LINT_ANDROIDX_MISC Category
Historically, one use of the categories was to associate issues with a
link to online document for API Guidelines. AndroidX had a number of
different pages and so needed multiple categories. Since those links
have been removed it no longer needs a special AndroidX category so
this change removes it.

Bug: 244774444
Test: ./gradlew
Change-Id: I3796028d27773739001590f793a1521592d2be64