FlaggedApiTest: Make more realistic by using Flags.FLAG_FOO_BAR
Previously, the `FlaggedApiLiteral` check was suppressed for this test
because it used literal strings everywhere. This change removes that
suppression and fixes the issues that arose.

Unfortunately, it was necessary to disable these tests for Turbine
because it does not yet provide access to the source code of annotation
attributes values, only the literal value. There was also a slight name
change of the tests from `...foo/bar...` to `...foo_bar...` that caused
a little bit of churn in the baseline file.

Bug: 339794405
Test: ./gradlew
Change-Id: I91e8b47563556da05a72a594d13a960189469fa5