Add Severity.UNLIMITED for default value of maximumSeverity
Previously, the `maximumSeverity` parameter of the
`Reporter.report(...)` methods was `ERROR` which was the highest
`Severity`. This change adds a new `Severity.UNLIMITED` property which
is an alias of `Severity.ERROR` for use as the default value of
`maximumSeverity` parameter and related code. This makes it clearer
that it is actually the highest `Severity`.

Bug: 339166506
Test: ./gradlew
Change-Id: I0b241ec0e2a737e103b75a48b984b031940b6af5